feat: refactor to src-format and update dependencies for Python 3.11
- Move package to src/ directory following modern Python packaging practices
- Update Python requirement to >=3.11 and bump version to 1.5.0
- Update dependencies to newer versions and add click as required dependency
- Update build system requirements and configuration
- Add uv.lock file for dependency locking
- Modify Makefile to use python -m pytest for better test execution
- Reorganize dev dependencies under [dependency-groups] in pyproject.toml
